rename now ill-named rpubshtmlgenerator -> statichtmlgenerator
