+++ src/gwt/src/org/rstudio/studio/client/rsconnect/model/StaticHtmlGenerator.java
@@ -1,7 +1,7 @@
 /*
- * RPubsHtmlGenerator.java
+ * StaticHtmlGenerator.java
  *
- * Copyright (C) 2009-12 by RStudio, Inc.
+ * Copyright (C) 2009-15 by RStudio, Inc.
  *
  * Unless you have received this program directly from RStudio pursuant
  * to the terms of a commercial license agreement with RStudio, then
@@ -13,13 +13,13 @@
  *
  */
 
-package org.rstudio.studio.client.common.rpubs;
+package org.rstudio.studio.client.rsconnect.model;
 
 import org.rstudio.core.client.CommandWithArg;
 
-public interface RPubsHtmlGenerator
+public interface StaticHtmlGenerator
 {
-   void generateRPubsHtml(String title,
-                          String comment,
-                          CommandWithArg<String> onCompleted);
+   void generateStaticHtml(String title,
+                           String comment,
+                           CommandWithArg<String> onCompleted);
 }
